Update how Exchange and Teams interact
After facing with the issue first-hand, I suggest the following change to be made to the article, adding the following information, which would be extremely helpful imo, as Exchange hybrid is widely used with Teams and team channel meetings are a big thing nowadays:

If members of a channel, for which a Team meeting is being scheduled for, would like to receive email invitation for the meeting also (not only in-teams-channel notification), they would need to subscribe to the O365 group mailbox in their Inbox from Outlook. This feature is only supported from Exchange 2016 CU5 and above.